% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ar.R
\name{Ar}
\alias{Ar}
\title{Ar: Archimedes number}
\usage{
Ar(T_leaf, surface, pars)
}
\arguments{
\item{T_leaf}{Leaf temperature in Kelvin}

\item{surface}{Leaf surface (lower or upper)}

\item{pars}{Concatenated parameters (\code{leaf_par}, \code{enviro_par}, and \code{constants})}
}
\value{
A unitless number of class \code{units}
}
\description{
Ar: Archimedes number
}
\details{
The Archimedes number is a dimensionless number that describes when free or forced convection dominates. depends on a combination how much free or forced convection predominates.

\deqn{Ar = Gr / Re ^ 2}
\cr
\tabular{lllll}{
\emph{Symbol} \tab \emph{R} \tab \emph{Description} \tab \emph{Units} \tab \emph{Default}\cr
\eqn{Gr} \tab \code{Gr} \tab Grashof number \tab none \tab \link[=.get_gr]{calculated}\cr
\eqn{Re} \tab \code{Re} \tab Reynolds number \tab none \tab \link[=.get_re]{calculated}
}
}
\examples{
cs <- make_constants()
lp <- make_enviropar()
ep <- make_leafpar()
pars <- c(cs, lp, ep)
T_leaf <- set_units(298.15, "K")

Ar(T_leaf, "lower", pars)
Ar(T_leaf, "upper", pars)

}
