% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tleaves.R
\name{.get_dwv}
\alias{.get_dwv}
\title{d_wv: water vapour gradient (mol / m ^ 3)}
\usage{
.get_dwv(T_leaf, pars)
}
\arguments{
\item{T_leaf}{Leaf temperature in Kelvin}

\item{pars}{Concatenated parameters (\code{leaf_par}, \code{enviro_par}, and \code{constants})}
}
\value{
Value in mol / m^3 of class \code{units}
}
\description{
d_wv: water vapour gradient (mol / m ^ 3)
}
\details{
\bold{Water vapour gradient:} The water vapour pressure differential from inside to outside of the leaf is the saturation water vapor pressure inside the leaf (\eqn{p_\text{leaf}}{p_leaf}) minus the water vapor pressure of the air (\eqn{p_\text{air}}{p_air}):

\deqn{d_\text{wv} = p_\text{leaf} / (R T_\text{leaf}) - RH p_\text{air} / (R T_\text{air})}{d_wv = p_leaf / (R T_leaf) - RH p_air / (R T_air)}

Note that water vapor pressure is converted from kPa to mol / m^3 using ideal gas law. \cr

\tabular{lllll}{
\emph{Symbol} \tab \emph{R} \tab \emph{Description} \tab \emph{Units} \tab \emph{Default}\cr
\eqn{p_\text{air}}{p_air} \tab \code{p_air} \tab saturation water vapour pressure of air \tab kPa \tab \link[=.get_ps]{calculated}\cr
\eqn{p_\text{leaf}}{p_leaf} \tab \code{p_leaf} \tab saturation water vapour pressure inside the leaf \tab kPa \tab \link[=.get_ps]{calculated}\cr
\eqn{R} \tab \code{R} \tab ideal gas constant \tab J / (mol K) \tab 8.3144598\cr
\eqn{\text{RH}}{RH} \tab \code{RH} \tab relative humidity \tab \% \tab 0.50\cr
\eqn{T_\text{air}}{T_air} \tab \code{T_air} \tab air temperature \tab K \tab 298.15\cr
\eqn{T_\text{leaf}}{T_leaf} \tab \code{T_leaf} \tab leaf temperature \tab K \tab input
}
}
\examples{

# Water vapour gradient: 

leaf_par <- make_leafpar()
enviro_par <- make_enviropar()
constants <- make_constants()
pars <- c(leaf_par, enviro_par, constants)
T_leaf <- set_units(300, "K")
T_air <- set_units(298.15, "K")
p_leaf <- set_units(35.31683, "kPa")
p_air <- set_units(31.65367, "kPa")

d_wv <- p_leaf / (pars$R * T_leaf) - pars$RH * p_air / (pars$R * T_air)

}
