% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tealeaves.R
\name{.get_Pa}
\alias{.get_Pa}
\title{P_a: density of dry air (g / m^3)}
\usage{
.get_Pa(T_leaf, pars)
}
\arguments{
\item{T_leaf}{Leaf temperature in Kelvin}

\item{pars}{Concatenated parameters (\code{leaf_par}, \code{enviro_par}, and \code{constants})}
}
\value{
Value in g / m\eqn{^3} of class \code{units}
}
\description{
P_a: density of dry air (g / m^3)
}
\details{
\deqn{P_\text{a} = P / (R_\text{air} (T_\text{leaf} - T_\text{air}) / 2)}{P_a = P / (R_air (T_leaf - T_air) / 2)}

\tabular{lllll}{
\emph{Symbol} \tab \emph{R} \tab \emph{Description} \tab \emph{Units} \tab \emph{Default}\cr
\eqn{P} \tab \code{P} \tab atmospheric pressure \tab kPa \tab 101.3246\cr
\eqn{R_\text{air}}{R_air} \tab \code{R_air} \tab specific gas constant for dry air \tab J / (kg K) \tab 287.058\cr
\eqn{T_\text{air}}{T_air} \tab \code{T_air} \tab air temperature \tab K \tab 298.15\cr
\eqn{T_\text{leaf}}{T_leaf} \tab \code{T_leaf} \tab leaf temperature \tab K \tab input
}
}
