% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tleaves.R
\name{.get_ps}
\alias{.get_ps}
\title{Saturation water vapour pressure (kPa)}
\usage{
.get_ps(Temp, P)
}
\arguments{
\item{Temp}{Temperature in Kelvin}

\item{P}{Atmospheric pressure in kPa}
}
\value{
Value in kPa of class \code{units}
}
\description{
Saturation water vapour pressure (kPa)
}
\details{
Goff-Gratch equation (see http://cires1.colorado.edu/~voemel/vp.html) \cr
\cr
This assumes P = 1 atm = 101.3246 kPa, otherwise boiling temperature needs to change \cr
\cr
This returns p_s in hPa, which is converted to kPa: \cr
\cr
\deqn{p_\text{s} = 10 ^ (-7.90298 * (373.16 / Temp - 1) + 5.02808 * log_{10}(373.16 / Temp) - 1.3816e-7 * (10 ^ (11.344 * (1 - Temp / 373.16) - 1)) + 8.1328e-3 * (10 ^ (-3.49149 * (373.16 / Temp - 1)) - 1) + log_{10}(P))}{p_s = 10 ^ (-7.90298 * (373.16 / Temp - 1) + 5.02808 * log10(373.16 / Temp) - 1.3816e-7 * (10 ^ (11.344 * (1 - Temp / 373.16) - 1)) + 8.1328e-3 * (10 ^ (-3.49149 * (373.16 / Temp - 1)) - 1) + log10(P))}
}
\references{
\url{http://cires1.colorado.edu/~voemel/vp.html}
}
