% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tealeaves.R
\name{.get_Tv}
\alias{.get_Tv}
\title{Calculate virtual temperature}
\usage{
.get_Tv(Temp, p_air, P)
}
\arguments{
\item{Temp}{Temperature in Kelvin}

\item{p_air}{Saturation water vapour pressure of air in kPa}

\item{P}{Atmospheric pressure in kPa}
}
\value{
Value in K of class \code{units}
}
\description{
Calculate virtual temperature
}
\details{
\deqn{T_\text{v} = T / [1 - 0.388 (p_\text{air} / P)]}{T_v = T / [1 - 0.388 (p_air / P)]}

\tabular{lllll}{
\emph{Symbol} \tab \emph{R} \tab \emph{Description} \tab \emph{Units} \tab \emph{Default}\cr
\eqn{p_\text{air}}{p_air} \tab \code{p_air} \tab saturation water vapour pressure of air \tab kPa \tab \link[=.get_ps]{calculated}\cr
\eqn{P} \tab \code{P} \tab atmospheric pressure \tab kPa \tab 101.3246
}
}
