% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tealeaves.R
\name{engery_balance}
\alias{engery_balance}
\title{Calculate leaf energy balance}
\usage{
engery_balance(T_leaf, leaf_par, enviro_par, constants, quiet = FALSE,
  abs_val = FALSE)
}
\arguments{
\item{T_leaf}{Leaf temperature in Kelvin. If input is numeric, it will be automatically converted to \code{units}.}

\item{leaf_par}{A list of leaf parameters. This can be generated using the \code{make_leafpar} function.}

\item{enviro_par}{A list of environmental parameters. This can be generated using the \code{make_enviropar} function.}

\item{constants}{A list of physical constants. This can be generated using the \code{make_constants} function.}

\item{quiet}{Logical. Should a message appear about conversion from \code{numeric} to \code{units}?}

\item{abs_val}{Return absolute value? Useful for finding leaf temperature that balances heat transfer.}
}
\description{
Calculate leaf energy balance
}
