% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tealeaves-package.R
\docType{package}
\name{tealeaves}
\alias{tealeaves}
\alias{tealeaves-package}
\title{\code{tealeaves} package}
\description{
Solve for leaf temperature using energy balance
}
\details{
See the README on
\href{https://github.com/cdmuir/leafevolver}{GitHub}
}
