% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tl_example1}
\alias{tl_example1}
\title{{tealeaves} example output 1}
\format{A data frame with 150 rows and 20 variables:}
\usage{
tl_example1
}
\description{
An example output from the \code{\link{tealeaves}} function.
}
\keyword{datasets}
