% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tealeaves.R
\name{tleaf}
\alias{tleaf}
\title{Find leaf temperature}
\usage{
tleaf(leaf_par, enviro_par, constants)
}
\arguments{
\item{leaf_par}{A list of leaf parameters. This can be generated using the \code{make_leafpar} function.}

\item{enviro_par}{A list of environmental parameters. This can be generated using the \code{make_enviropar} function.}

\item{constants}{A list of physical constants. This can be generated using the \code{make_constants} function.}

\item{...}{Additional arguments passed to \code{optim}}
}
\value{
An object of class \code{tealeaf}
}
\description{
Find leaf temperature
}
\examples{

leaf_par <- make_leafpar()
enviro_par <- make_enviropar()
constants <- make_constants()
pars <- c(leaf_par, enviro_par, constants)
T_leaf <- tleaf(leaf_par, enviro_par, constants)
T_leaf

}
