% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tleaves.R
\name{tleaves}
\alias{tleaves}
\alias{tleaf}
\title{\code{tleaves}: find leaf temperatures for multiple parameter sets}
\usage{
tleaves(leaf_par, enviro_par, constants, n_start = 1, progress = TRUE,
  quiet = FALSE)

tleaf(leaf_par, enviro_par, constants, n_start = 1, progress = TRUE,
  quiet = FALSE)
}
\arguments{
\item{leaf_par}{A list of leaf parameters. This can be generated using the \code{make_leafpar} function.}

\item{enviro_par}{A list of environmental parameters. This can be generated using the \code{make_enviropar} function.}

\item{constants}{A list of physical constants. This can be generated using the \code{make_constants} function.}

\item{n_start}{Number of initial values to try in \code{\link[stats]{optim}}. The default is \code{n_start = 1} with the initial guess being \code{enviro_par$T_air}. If \code{n_start > 1}, the function will try \code{n_start} evenly spaced initial guesses between \code{enviro_par$T_air +/- 30}. Greater \code{n_start} will slow down optimization.}

\item{progress}{Logical. Should a progress bar be displayed?}

\item{quiet}{Logical. Should messages be displayed?}
}
\value{
\code{tleaves}: \cr
\cr

\cr
\code{tleaf}: \cr
\cr
A data.frame with the following numeric columns: \cr

\tabular{ll}{
\code{T_leaf} \tab Equilibrium leaf tempearture (K) \cr
\code{init} \tab Initial leaf tempearture (K) for optim search \cr
\code{value} \tab Leaf energy balance (W / m^2) at \code{tleaf} \cr
\code{convergence} \tab Convergence code (see \code{\link[stats]{optim}}) \cr
\code{R_abs} \tab Total absorbed radiation (W / m^2; see \code{\link{.get_Rabs}}) \cr
\code{S_r} \tab Longwave re-radiation (W / m^2; see \code{\link{.get_Sr}}) \cr
\code{H} \tab Sensible heat flux density (W / m^2; see \code{\link{.get_H}}) \cr
\code{L}\tab Latent heat flux density (W / m^2; see \code{\link{.get_L}})
}
}
\description{
\code{tleaves}: find leaf temperatures for multiple parameter sets

\code{tleaf}: find leaf temperatures for a single parameter set
}
\examples{

leaf_par <- make_leafpar()
enviro_par <- make_enviropar()
constants <- make_constants()
pars <- c(leaf_par, enviro_par, constants)
T_leaf <- tleaf(leaf_par, enviro_par, constants)
T_leaf

}
