% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tealeaves.R
\docType{package}
\name{tealeaves}
\alias{tealeaves}
\alias{tealeaves-package}
\title{\code{tealeaves} package}
\description{
Solve for Leaf Temperature Using Energy Balance
}
\details{
See the README on
\href{https://github.com/cdmuir/tealeaves}{GitHub}
}
